wallet_service/requirements.txt
asgiref==3.3.1
coverage==5.5
Django==3.1.7
djangorestframework==3.12.2
nose==1.3.7
psycopg2-binary==2.8.6
pycodestyle==2.7.0
pytz==2021.1
sqlparse==0.4.1
